requirements.txt
bleson~=0.1.8
paho-mqtt~=2.1.0